backend/requirements.txt
Flask==2.2.2
Flask-SocketIO==5.5.1
eventlet==0.40.0
Flask-Bcrypt==1.0.1
pymongo==4.3.3
python-dotenv==1.0.0
Werkzeug==2.2.3
Flask-Cors==4.0.1
